# perf(xamlgen): Reduce allocations in RewriteNamespaces