closes #14527

Benchmark:

```csharp
    [Params("Abc", "muxc:Abc", "(muxc:Abc)", "M(muxc:Abc.A, muxc:Abc.B)")]
    public string xamlString { get; set; }

    [Params("muxc")]
    public string nsPrefix { get; set; }

    [Params("using:Microsoft.UI.Xaml.Controls")]
    public string nsNamespace { get; set; }


    [Benchmark(Baseline = true)]
    public string RegexBench()
        => Regex.Replace(xamlString, $@"(^|[^\w])({nsPrefix}:)", "$1global::" + nsNamespace.TrimStart("using:") + ".");

    [Benchmark]
    public string WithSpanSlicingAndCopying()
    {
        if (nsNamespace.StartsWith("using:", StringComparison.Ordinal))
        {
            while (xamlString.Length > nsPrefix.Length)
            {
                var index = xamlString.IndexOf($"{nsPrefix}:", StringComparison.Ordinal);
                if (index == 0 || (index > 0 && !char.IsLetterOrDigit(xamlString[index - 1])))
                {
                    var nsGlobalized = nsNamespace.Replace("using:", "global::");

                    // following is the equivalent of "string.Concat(xamlString.Substring(0, index), nsGlobalized, ".", xamlString.Substring(index + ns.Prefix.Length + 1))"
                    // but in a way that allocates less memory.
                    var newLength = xamlString.Length + nsGlobalized.Length - nsPrefix.Length;
                    xamlString = StringExtensions.Create(newLength, (xamlString, index, nsGlobalized, nsPrefix), static (span, state) =>
                    {
                        var (xamlString, index, nsGlobalized, nsPrefix) = state;
                        var copiedLengthSoFar = 0;

                        xamlString.AsSpan().Slice(0, index).CopyTo(span.Slice(copiedLengthSoFar));
                        copiedLengthSoFar += index;

                        nsGlobalized.AsSpan().CopyTo(span.Slice(copiedLengthSoFar));
                        copiedLengthSoFar += nsGlobalized.Length;

                        span[copiedLengthSoFar] = '.';
                        copiedLengthSoFar++;

                        xamlString.AsSpan().Slice(index + nsPrefix.Length + 1).CopyTo(span.Slice(copiedLengthSoFar));
                    });
                }
                else
                    break;
            }

            return xamlString;
        }
        else
        {
            return xamlString;
        }
    }
```

and

```csharp
public static class Extensions
{
    public static string TrimStart(this string source, string trimText, bool ignoreCase = false)
    {
        if (!string.IsNullOrEmpty(trimText) && source.StartsWith(trimText, ignoreCase ? StringComparison.OrdinalIgnoreCase : StringComparison.Ordinal))
        {
            return source.Substring(trimText.Length);
        }
        else
        {
            return source;
        }
    }
}
```

and

```csharp
public delegate void SpanAction<T, in TArg>(Span<T> span, TArg arg);


namespace MyBench
{
    public static class StringExtensions
    {

        public static string Create<TState>(int length, TState state, SpanAction<char, TState> action)
        {
            scoped Span<char> buffer;

            if (length <= 128)
                buffer = stackalloc char[512];
            else
                buffer = new char[length];

            action(buffer.Slice(0, length), state);
            return buffer.Slice(0, length).ToString();
        }
    }
}
```

Result:

|                    Method |           xamlString | nsPrefix |          nsNamespace |         Mean |     Error |    StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------- |--------------------- |--------- |--------------------- |-------------:|----------:|----------:|------:|-------:|------:|------:|----------:|
|                RegexBench |           (muxc:Abc) |     muxc | using(...)trols [32] |   393.654 ns | 1.5770 ns | 1.3168 ns |  1.00 | 0.1602 |     - |     - |     336 B |
| WithSpanSlicingAndCopying |           (muxc:Abc) |     muxc | using(...)trols [32] |    18.640 ns | 0.0729 ns | 0.0647 ns |  0.05 | 0.0153 |     - |     - |      32 B |
|                           |                      |          |                      |              |           |           |       |        |       |       |           |
|                RegexBench |                  Abc |     muxc | using(...)trols [32] |   118.172 ns | 0.3048 ns | 0.2545 ns |  1.00 | 0.1109 |     - |     - |     232 B |
| WithSpanSlicingAndCopying |                  Abc |     muxc | using(...)trols [32] |     2.142 ns | 0.0088 ns | 0.0082 ns |  0.02 |      - |     - |     - |         - |
|                           |                      |          |                      |              |           |           |       |        |       |       |           |
|                RegexBench | M(mux(...)bc.B) [25] |     muxc | using(...)trols [32] | 1,092.265 ns | 3.1872 ns | 2.9813 ns |  1.00 | 0.2022 |     - |     - |     424 B |
| WithSpanSlicingAndCopying | M(mux(...)bc.B) [25] |     muxc | using(...)trols [32] |    20.957 ns | 0.1646 ns | 0.1540 ns |  0.02 | 0.0153 |     - |     - |      32 B |
|                           |                      |          |                      |              |           |           |       |        |       |       |           |
|                RegexBench |             muxc:Abc |     muxc | using(...)trols [32] |   330.517 ns | 0.9944 ns | 0.9301 ns |  1.00 | 0.1602 |     - |     - |     336 B |
| WithSpanSlicingAndCopying |             muxc:Abc |     muxc | using(...)trols [32] |    18.535 ns | 0.0431 ns | 0.0403 ns |  0.06 | 0.0153 |     - |     - |      32 B |

That is A MUCH more improvement than what I was expecting.